RTH High Level APIs (#155)
The primitives for Recoverable Transaction History - https://github.com/mobilecoinfoundation/mcips/blob/main/text/0004-recoverable-transaction-history.md - have already been implemented, but making it easy to use for SDK consumers requires some additional work. The high-level API accepts a list of "contacts" (or a single contact) and optionally a list of transactions. It returns a sorted "history" of transactions with memo's recovered if we're able to validate them, and the associated contact, if its a match.

* Public High-Level APIs that return a recovered transaction history
* Contact-like protocol PublicAddressProvider that consumers can use to make contacts
* Unit tests